src/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pydantic
sqlparse
requests
gunicorn
pytest
python-dotenv
langchain-google-genai
pytest
pytest-asyncio
httpx